requirements.txt
aiofiles==0.4.0
beautifulsoup4==4.8.2
dash==1.11.0
fake-useragent==0.1.11
fastapi==0.54.0
fasttext==0.9.2
gunicorn==20.0.4
gevent==1.4.0
google-cloud-storage>=1.28.1
html2text==2019.8.11
joblib==0.14.1
matplotlib==2.2.2
mypy==0.770
mypy-extensions==0.4.3
nlpaug==0.0.13.3
nltk==3.4.5
pandas==1.0.1
pycld3==0.20
PySnooper==0.0.33
python-dotenv>=0.10.1
requests==2.22.0
scikit-learn==0.22.2.post1
setuptools>=39.1.0
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
-f https://download.pytorch.org/whl/torch_stable.html                                                                                                                                                             
torch==1.4.0+cpu
torchvision==0.5.0+cpu
transformers==2.5.1
ujson==2.0.3
uvicorn==0.11.3
